[clang-tidy] An option for conditional skipping overloaded functions in modernize-use-string-view
