+++ clang-tools-extra/clang-tidy/modernize/UseStringViewCheck.cpp
@@ -22,7 +22,9 @@ using namespace clang::ast_matchers;
 namespace clang::tidy::modernize {
 
 namespace {
-AST_MATCHER(FunctionDecl, isOverloaded) {
+AST_MATCHER_P(FunctionDecl, isOverloaded, bool, CheckOverloadedFunctions) {
+  if (CheckOverloadedFunctions)
+    return false;
   const DeclarationName Name = Node.getDeclName();
   // Sanity check
   if (Name.isEmpty())
@@ -81,7 +83,8 @@ UseStringViewCheck::UseStringViewCheck(StringRef Name,
                                        ClangTidyContext *Context)
     : ClangTidyCheck(Name, Context),
       IgnoredFunctions(utils::options::parseStringList(
-          Options.get("IgnoredFunctions", "toString$;ToString$;to_string$"))) {
+          Options.get("IgnoredFunctions", "toString$;ToString$;to_string$"))),
+      CheckOverloadedFunctions(Options.get("CheckOverloadedFunctions", false)) {
   parseReplacementStringViewClass(
       Options.get("ReplacementStringViewClass", ""));
 }
@@ -89,6 +92,7 @@ UseStringViewCheck::UseStringViewCheck(StringRef Name,
 void UseStringViewCheck::storeOptions(ClangTidyOptions::OptionMap &Opts) {
   Options.store(Opts, "IgnoredFunctions",
                 utils::options::serializeStringList(IgnoredFunctions));
+  Options.store(Opts, "CheckOverloadedFunctions", CheckOverloadedFunctions);
   Options.store(Opts, "ReplacementStringViewClass",
                 (Twine("") + StringViewClassKey + "=" + StringViewClass + ";" +
                  WStringViewClassKey + "=" + WStringViewClass + ";" +
@@ -113,7 +117,7 @@ void UseStringViewCheck::registerMatchers(MatchFinder *Finder) {
       functionDecl(
           isDefinition(),
           unless(anyOf(VirtualOrOperator, IgnoredFunctionsMatcher,
-                       isOverloaded(),
+                       isOverloaded(CheckOverloadedFunctions),
                        ast_matchers::isExplicitTemplateSpecialization())),
           returns(IsStdString), hasDescendant(returnStmt()),
           unless(hasDescendant(returnStmt(hasReturnValue(unless(
